+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 0.6.0
+## 1.1.0
+  * Updated python version 3.12 [#30](https://github.com/singer-io/tap-codat/pull/30)
+  * Added integration tests.
+
+## 1.0.0
+  * State bookmark structure changed from `bookmarks[stream][companyId]` to per-company keying under each stream. [#29](https://github.com/singer-io/tap-codat/pull/29)
   * Update libraries [#26](https://github.com/singer-io/tap-codat/pull/26)
   * Bug fix: NoneType object has no attribute lower